requirements.txt
python-dotenv
langchain-classic
langchain-text-splitters
langchain-community
langchain-google-genai
chromadb
pypdf
google-generativeai
